refactor - move catalog
